Sentraliserer egne Type deklarasjoner i /types slik at det blir lettere å lese komponentene
