+++ src/js/components/meldinger/MinInnboks.js
@@ -1,11 +1,11 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import { FormattedMessage as F, injectIntl, intlShape } from 'react-intl';
 import i18n from '../../../translations/i18n';
 import LenkepanelMedIkon from '../common/LenkepanelMedIkon';
 import PanelOverskrift from '../common/PanelOverskrift';
 import IkonOppgave from '../../../assets/IkonOppgave';
 import IkonInnboks from '../../../assets/IkonInnboks';
+import MinInnboksType from '../../types/MinInnboksType';
 
 const getMinInnboksIcon = (type) => {
   switch (type) {
@@ -55,11 +55,6 @@ const MinInnboks = ({ mininnboks, intl }) => {
     </>
   );
 };
-
-export const MinInnboksType = PropTypes.arrayOf(PropTypes.shape({
-  type: PropTypes.string.isRequired,
-  url: PropTypes.string,
-}));
 
 MinInnboks.propTypes = {
   mininnboks: MinInnboksType,
