+++ src/js/types/HendelserType.js
@@ -1,7 +1,13 @@
 import PropTypes from 'prop-types';
 
 const HendelserType = PropTypes.arrayOf(
-  PropTypes.oneOfType([PropTypes.object]),
+  PropTypes.shape({
+    uid: PropTypes.string,
+    eventId: PropTypes.string.isRequired,
+    type: PropTypes.string.isRequired,
+    tekst: PropTypes.string.isRequired,
+    link: PropTypes.string,
+  }),
 );
 
 export default HendelserType;
