+++ src/js/types/MinInnboksType.js
@@ -1,0 +1,8 @@
+import PropTypes from 'prop-types';
+
+const MinInnboksType = PropTypes.arrayOf(PropTypes.shape({
+  type: PropTypes.string.isRequired,
+  url: PropTypes.string,
+}));
+
+export default MinInnboksType;
