+++ src/js/types/PaabegynteSoknaderType.js
@@ -1,0 +1,8 @@
+import PropTypes from 'prop-types';
+
+const PaabegynteSoknaderType = PropTypes.shape({
+  url: PropTypes.string.isRequired,
+  antallPaabegynte: PropTypes.number.isRequired,
+});
+
+export default PaabegynteSoknaderType;
